requirements.txt
python-telegram-bot
yt-dlp
fastapi
uvicorn
